# Conflicts:
#	airflow/providers/amazon/aws/transfers/redshift_to_s3.py
#	tests/providers/amazon/aws/transfers/test_redshift_to_s3.py

+++ airflow/migrations/versions/cf5dc11e79ad_drop_user_and_chart.py
@@ -18,7 +18,7 @@
 """drop_user_and_chart
 
 Revision ID: cf5dc11e79ad
-Revises: 41f5f12752f8
+Revises: 03afc6b6f902
 Create Date: 2019-01-24 15:30:35.834740
 
 """
@@ -29,12 +29,12 @@ from sqlalchemy.engine.reflection import Inspector
 
 # revision identifiers, used by Alembic.
 revision = 'cf5dc11e79ad'
-down_revision = '41f5f12752f8'
+down_revision = '03afc6b6f902'
 branch_labels = None
 depends_on = None
 
 
-def upgrade():
+def upgrade():  # noqa: D103
     # We previously had a KnownEvent's table, but we deleted the table without
     # a down migration to remove it (so we didn't delete anyone's data if they
     # were happing to use the feature.
@@ -43,15 +43,21 @@ def upgrade():
 
     conn = op.get_bind()
     inspector = Inspector.from_engine(conn)
+    tables = inspector.get_table_names()
 
-    if 'known_event' in inspector.get_table_names() != 'sqlite':
+    if 'known_event' in tables:
         op.drop_constraint('known_event_user_id_fkey', 'known_event')
 
-    op.drop_table("chart")
-    op.drop_table("users")
+    if "chart" in tables:
+        op.drop_table(
+            "chart",
+        )
+
+    if "users" in tables:
+        op.drop_table("users")
 
 
-def downgrade():
+def downgrade():  # noqa: D103
     conn = op.get_bind()
 
     op.create_table(
@@ -62,7 +68,7 @@ def downgrade():
         sa.Column('password', sa.String(255)),
         sa.Column('superuser', sa.Boolean(), default=False),
         sa.PrimaryKeyConstraint('id'),
-        sa.UniqueConstraint('username')
+        sa.UniqueConstraint('username'),
     )
 
     op.create_table(
@@ -82,8 +88,11 @@ def downgrade():
         sa.Column('x_is_date', sa.Boolean(), nullable=True),
         sa.Column('iteration_no', sa.Integer(), nullable=True),
         sa.Column('last_modified', sa.DateTime(), nullable=True),
-        sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
-        sa.PrimaryKeyConstraint('id')
+        sa.ForeignKeyConstraint(
+            ['user_id'],
+            ['users.id'],
+        ),
+        sa.PrimaryKeyConstraint('id'),
     )
 
     if conn.dialect.name == 'mysql':
